Moved `Timer` to `utils/stdlib.py`
It used to be in `utils/testing.py`, which imported `pytest`, meaning
that any machine that wanted to use `Timer` needed `pytest`. Since
`Timer` can be used for much more than tests, it makes more sense
to have it in a generic location.
