+++ nengo/utils/stdlib.py
@@ -11,6 +11,7 @@ import itertools
 import os
 import shutil
 import sys
+import time
 
 from .compat import iteritems, reraise
 
@@ -167,3 +168,40 @@ def nested(*managers):
                 exc = sys.exc_info()
         if exc != (None, None, None):
             reraise(exc[0], exc[1], exc[2])
+
+
+class Timer(object):
+    """A context manager for timing a block of code.
+
+    Attributes
+    ----------
+    duration : float
+        The difference between the start and end time (in seconds).
+        Usually this is what you care about.
+    start : float
+        The time at which the timer started (in seconds).
+    end : float
+        The time at which the timer ended (in seconds).
+
+    Example
+    -------
+    >>> import time
+    >>> with Timer() as t:
+    ...    time.sleep(1)
+    >>> assert t.duration >= 1
+    """
+
+    TIMER = time.clock if sys.platform == "win32" else time.time
+
+    def __init__(self):
+        self.start = None
+        self.end = None
+        self.duration = None
+
+    def __enter__(self):
+        self.start = Timer.TIMER()
+        return self
+
+    def __exit__(self, type, value, traceback):
+        self.end = Timer.TIMER()
+        self.duration = self.end - self.start
